Rename GameManger to GameManager
typo
