+++ Assets/Scenes/DoorScene2.unity
@@ -405,48 +405,49 @@ Prefab:
     m_RemovedComponents: []
   m_ParentPrefab: {fileID: 100100000, guid: 48b70681550ab6341bfa47b47a369e7a, type: 2}
   m_IsPrefabParent: 0
---- !u!1001 &2018894740
-Prefab:
+--- !u!1 &2058188963
+GameObject:
   m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 4379899491509790, guid: fedcf59c8b5603d448e6a96d48ed7253, type: 2}
-      propertyPath: m_LocalPosition.x
-      value: 0.013197534
-      objectReference: {fileID: 0}
-    - target: {fileID: 4379899491509790, guid: fedcf59c8b5603d448e6a96d48ed7253, type: 2}
-      propertyPath: m_LocalPosition.y
-      value: 0.15905507
-      objectReference: {fileID: 0}
-    - target: {fileID: 4379899491509790, guid: fedcf59c8b5603d448e6a96d48ed7253, type: 2}
-      propertyPath: m_LocalPosition.z
-      value: -0.012316623
-      objectReference: {fileID: 0}
-    - target: {fileID: 4379899491509790, guid: fedcf59c8b5603d448e6a96d48ed7253, type: 2}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 4379899491509790, guid: fedcf59c8b5603d448e6a96d48ed7253, type: 2}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 4379899491509790, guid: fedcf59c8b5603d448e6a96d48ed7253, type: 2}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 4379899491509790, guid: fedcf59c8b5603d448e6a96d48ed7253, type: 2}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 4379899491509790, guid: fedcf59c8b5603d448e6a96d48ed7253, type: 2}
-      propertyPath: m_RootOrder
-      value: 5
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_ParentPrefab: {fileID: 100100000, guid: fedcf59c8b5603d448e6a96d48ed7253, type: 2}
-  m_IsPrefabParent: 0
+  m_PrefabParentObject: {fileID: 1727541477528204, guid: fedcf59c8b5603d448e6a96d48ed7253,
+    type: 2}
+  m_PrefabInternal: {fileID: 0}
+  serializedVersion: 5
+  m_Component:
+  - component: {fileID: 2058188965}
+  - component: {fileID: 2058188964}
+  m_Layer: 0
+  m_Name: _GM
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!114 &2058188964
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_PrefabParentObject: {fileID: 114715530406952822, guid: fedcf59c8b5603d448e6a96d48ed7253,
+    type: 2}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 2058188963}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 4f5cd1e8d1e16414e9662937a526ca6d, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!4 &2058188965
+Transform:
+  m_ObjectHideFlags: 0
+  m_PrefabParentObject: {fileID: 4379899491509790, guid: fedcf59c8b5603d448e6a96d48ed7253,
+    type: 2}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 2058188963}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0.013197534, y: 0.15905507, z: -0.012316623}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_RootOrder: 5
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1001 &2079845841
 Prefab:
   m_ObjectHideFlags: 0
